@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "A CLI command to build a simple Vue.js SPA with almost no configuration.",
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "bin": "bin/index.js",
   "main": "lib/index.js",
@@ -12,33 +12,62 @@
     "lib"
   ],
   "dependencies": {
-    "autoprefixer": "^6.7.6",
-    "babel-core": "^6.23.1",
+    "assert": "^1.4.1",
+    "autoprefixer": "^7.1.1",
+    "babel-core": "^6.24.0",
     "babel-helper-vue-jsx-merge-props": "^2.0.2",
+    "babel-plugin-istanbul": "^4.1.1",
+    "babel-plugin-minify-constant-folding": "^0.1.3",
+    "babel-plugin-minify-dead-code-elimination": "^0.1.4",
     "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-async-to-generator": "^6.22.0",
-    "babel-plugin-transform-inline-environment-variables": "0.0.2",
-    "babel-plugin-transform-vue-jsx": "^3.3.0",
+    "babel-plugin-transform-inline-environment-variables": "^0.1.1",
+    "babel-plugin-transform-vue-jsx": "^3.4.2",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-babili": "0.0.12",
-    "babel-preset-es2015": "^6.22.0",
-    "babelify": "^7.3.0",
-    "browserify": "^14.1.0",
-    "cpx": "^1.5.0",
+    "babel-preset-babili": "^0.1.4",
+    "babel-preset-env": "^1.3.2",
+    "chalk": "^1.1.3",
+    "chokidar": "^1.6.1",
+    "cross-spawn": "^5.1.0",
+    "cssnano": "^3.10.0",
+    "dot-prop": "^4.1.1",
+    "fs-promise": "^2.0.2",
+    "karma": "^1.5.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-coverage": "^1.1.1",
+    "karma-detect-browsers": "^2.2.5",
+    "karma-edge-launcher": "^0.4.1",
+    "karma-firefox-launcher": "^1.0.1",
+    "karma-growl-reporter": "^1.0.0",
+    "karma-ie-launcher": "^1.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-safari-launcher": "^1.0.0",
+    "magic-string": "^0.21.3",
     "mime": "^1.3.4",
     "minimist": "^1.2.0",
-    "postcss": "^5.2.16",
-    "postcss-import": "^9.1.0",
-    "vueify": "mysticatea/vueify",
-    "watchify": "^3.9.0"
+    "mocha": "^3.2.0",
+    "parse5": "^3.0.2",
+    "postcss": "^6.0.2",
+    "postcss-calc": "^6.0.0",
+    "postcss-custom-properties": "^6.0.1",
+    "postcss-import": "^10.0.0",
+    "read-pkg-up": "^2.0.0",
+    "rollup": "^0.43.0",
+    "rollup-plugin-alias": "^1.3.0",
+    "rollup-plugin-babel": "^2.7.1",
+    "rollup-plugin-commonjs": "^8.0.2",
+    "rollup-plugin-json": "^2.1.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-vue": "github:mysticatea/rollup-plugin-vue",
+    "util": "^0.10.3",
+    "vue": "^2.2.6"
   },
   "devDependencies": {
-    "eslint": "^3.17.1",
-    "eslint-config-mysticatea": "^9.0.2",
-    "material-design-icons": "3.0.1",
-    "mocha": "^3.2.0",
-    "shelljs": "^0.7.7",
-    "vue": "2.2.2"
+    "eslint": "^3.19.0",
+    "eslint-config-mysticatea": "^10.0.0",
+    "shelljs": "^0.7.7"
+  },
+  "peerDependencies": {
+    "babel-polyfill": "^6.23.0"
   },
   "scripts": {
     "lint": "eslint bin lib test",
